Add wind functionality
We are interested in finding out what is different (if anything) if we use wind generators vs solar generators. So, we add the functions that allow us to that.

In addition, we are finding that these generators are not really added--instead, we see that it prefers grid. However, we still want to do experiments, so we remove the grid and add functionality to investigate year, among other things, with just wind and just solar generators

That's the last thing, in that it is confusing to use 'nogrid' or 'nosolar' now. So instead, we are using 'justsolar', 'justwind', etc. This is not done everywhere for all the plotting functions, however